+++ packages/ts-lib-tools/src/gen-module.ts
@@ -9,6 +9,18 @@ import execa = require('execa');
 import ora = require('ora');
 
 /**
+ * 生成许可证文件
+ *
+ * @param modulePath 模块路径
+ */
+async function genLicenseFile(modulePath: string) {
+  const isExists = await pathExists(resolveRoot('LICENSE'));
+  if (isExists) {
+    await copy(resolveRoot('LICENSE'), resolve(modulePath, 'LICENSE'));
+  }
+}
+
+/**
  * 生成模块
  *
  * @param {string} moduleName
@@ -32,7 +44,9 @@ async function genModule(moduleName: string) {
   // 更新package.json
   const packageInfo = await readJSON(resolve(modulePath, 'package.json'));
   const { version = '0.1.0' } = await readJSON(resolveRoot('lerna.json'));
-  const { devDependencies } = await readJSON(resolveRoot('package.json'));
+  const { devDependencies, license = 'MIT' } = await readJSON(
+    resolveRoot('package.json'),
+  );
   const { npmScope } = await readJSON(resolveRoot('ts-lib.config.json'));
   const name = `@${npmScope}/${moduleName}`;
   packageInfo.name = name;
@@ -40,9 +54,12 @@ async function genModule(moduleName: string) {
   packageInfo.devDependencies['ts-lib-tools'] = devDependencies['ts-lib-tools'];
   packageInfo['umd:main'] = `dist/${safePackageName(name)}.umd.production.js`;
   packageInfo.module = `dist/${safePackageName(name)}.es.production.js`;
+  packageInfo.license = license;
   await outputJSON(resolve(modulePath, 'package.json'), packageInfo, {
     spaces: 2,
   });
+
+  await genLicenseFile(modulePath);
 
   spinner.succeed(`已生成模块 ${chalk.green(moduleName)}`);