+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat: 规范 vscode 的换行符和缩进
 - fix(ts-lib-tools): 修复在 monorepo 模式下执行 jest 单元测试失败的缺陷
 - fix(ts-lib-tools): 修复 monorepo 模式下，yarn lint 找不到文件的缺陷
+- fix(ts-lib-scripts): 修复 monorepo 模式，lint-staged 配置错误
 
 ## v0.3.5